fixed all conflicts and removed svgs
