+++ frontend/control-center/src/components/ChannelAvatar/index.module.scss
