+++ frontend/control-center/src/components/Sidebar/index.tsx
@@ -17,22 +17,22 @@ export const Sidebar = () => {
   return (
     <nav className={styles.wrapper}>
       <div className={styles.linkSection}>
+        <div className={`${styles.align} ${isActive(COMPONENTS_ROUTE) ? styles.active : ''}`}>
+          <Link to={COMPONENTS_ROUTE} className={`${styles.link} ${isActive(COMPONENTS_ROUTE) ? styles.active : ''}`}>
+            <ComponentsIcon width={'24px'} height={'24px'} />
+            <span className={styles.iconText}>Components</span>
+          </Link>
+        </div>
         <div className={`${styles.align} ${isActive(CHANNELS_ROUTE) ? styles.active : ''}`}>
           <Link to={CHANNELS_ROUTE} className={`${styles.link} ${isActive(CHANNELS_ROUTE) ? styles.active : ''}`}>
             <ConnectorsIcon width={'24px'} height={'24px'} />
             <span className={styles.iconText}>Connectors</span>
           </Link>
         </div>
-        <div className={`${styles.align} ${isActive(CHANNELS_ROUTE) ? styles.active : ''}`}>
+        <div className={`${styles.align} ${isActive(CATALOG_ROUTE) ? styles.active : ''}`}>
           <Link to={CATALOG_ROUTE} className={`${styles.link} ${isActive(CATALOG_ROUTE) ? styles.active : ''}`}>
             <CatalogIcon width={'24px'} height={'24px'} />
-            <span className={styles.iconText}>Connectors</span>
-          </Link>
-        </div>
-        <div className={`${styles.align} ${isActive(CHANNELS_ROUTE) ? styles.active : ''}`}>
-          <Link to={COMPONENTS_ROUTE} className={`${styles.link} ${isActive(COMPONENTS_ROUTE) ? styles.active : ''}`}>
-            <ComponentsIcon width={'24px'} height={'24px'} />
-            <span className={styles.iconText}>Connectors</span>
+            <span className={styles.iconText}>Catalog</span>
           </Link>
         </div>
       </div>
