+++ frontend/control-center/src/pages/Channels/index.module.scss
@@ -5,11 +5,7 @@
   background: white;
   border-radius: 10px;
   padding: 32px;
-<<<<<<< HEAD:frontend/control-center/src/pages/Channels/index.module.scss
   margin: 88px 1.5em 0 191px;
-=======
-  margin: 88px 1.5em 0 96px;
->>>>>>> b22fef6b (moved channels from UI to control-center):frontend/control-center/src/pages/Channels/MainPage/index.module.scss
   height: calc(100vh - 88px);
   overflow-y: scroll;
   overflow-x: hidden;
